+++ server-side/src/main/java/com/codename_vp/serverside/Repository/OwnedListRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.codename_vp.serverside.Entity.OwnedList;
 
-public interface OwnedListRepo extends JpaRepository<OwnedList, Long> {
+public interface OwnedListRepo extends JpaRepository<OwnedList, Integer> {
     Optional<OwnedList> findByName(String name);
 
     Optional<OwnedList> removeByName(String name);
